requirements.txt
absl-py==0.8.0
appdirs==1.4.4
astor==0.8.1
attrs==19.3.0
bcolz==1.2.1
blis==0.2.4
bokeh==1.4.0
certifi==2020.4.5.1
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
ConfigArgParse==1.0
cycler==0.10.0
cymem==2.0.2
cytoolz==0.10.1
dask==2.11.0
decorator==4.4.2
distlib==0.3.0
distributed==2.11.0
easydict==1.9
filelock==3.0.12
Flask==1.1.1
fsspec==0.6.2
future==0.18.2
gast==0.3.3
google-pasta==0.2.0
GPUtil==1.4.0
graphviz==0.13.2
grpcio==1.23.0
gym==0.15.7
h5py==2.10.0
HeapDict==1.0.1
idna==2.8
imageio==2.8.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
locket==0.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
mock==4.0.1
more-itertools==8.3.0
msgpack==0.6.1
murmurhash==1.0.2
networkx==2.3
nltk==3.4.5
numexpr==2.7.1
numpy==1.16.4
olefile==0.46
opencv-python==4.2.0.34
OpenNMT-py==1.0.0
packaging==20.1
pandas==0.25.3
partd==1.1.0
Pillow==6.1.0
plac==0.9.6
pluggy==0.13.1
portalocker==1.5.1
preshed==2.0.1
protobuf==3.9.1
psutil==5.6.7
py==1.8.1
pycparser==2.19
pygame==1.9.6
pyglet==1.5.0
pyonmttok==1.18.1
pyparsing==2.4.6
pytest==5.4.2
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
requests==2.22.0
rouge==0.3.2
sacrebleu==1.3.7
scikit-image==0.16.2
scipy==1.4.1
six==1.12.0
sortedcontainers==2.1.0
spacy==2.1.8
srsly==0.1.0
tables==3.6.1
tblib==1.6.0
tensorboard==1.14.0
tensorboardX==1.8
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
thinc==7.0.8
toolz==0.10.0
torch==1.2.0
torchtext==0.4.0
torchvision==0.4.0a0+6b959ee
torchviz==0.0.1
tornado==6.0.3
tqdm==4.35.0
typing==3.7.4.1
urllib3==1.25.3
virtualenv==20.0.21
waitress==1.4.2
wasabi==0.2.2
wcwidth==0.1.9
Werkzeug==0.15.5
wrapt==1.12.1
zict==1.0.0
zipp==3.1.0
